ECR compatibility for /check
well, that was a nightmare. hours lost to that missing trailing slash,
and then a bit more time spent dealing with ECR's
non-conformant/out-of-date API for getting the digest, and dealing with
Docker's love of global state (hence the two underscored imports for
both schemas to register their unmarshal funcs) and some spent on that
global state becoming local thanks to vendor/.

ugh.

but it works now.