Add default to SKIP_PRIVILEGED
In the event that SKIP_PRIVILEGED is not set and common.sh is sourced
into a file that has `set -u`, it will error:

```
/opt/resource/common.sh: line 47: SKIP_PRIVILEGED: unbound variable
```

I discovered this while trying to build HEAD of docker-image-resource to
be used as a resource_type. The image built but put was failing with the
above error.

The test used will substitute parameter if it is set and not null,
otherwise set null if the parameter is null, otherwise use "" (null) if
it is unset. See http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_06_02 for a table explaining each.