Add support for using zfs
This adds the alpine zfs package so that docker can use the zfs
storage backend if it's available.

Additionally, this purges docker entirely before exitting to prevent
any resource leakage (something that, no surprise, happens when
you use the zfs storage backend nested. The zfs filesystems that the
inner docker creates are left as zombies on the host system)