do not start dockerd in 'in' script if skip_download is set to true
The 'in' script performs no Docker operations when `skip_download` is
set to true, so it does not make sense to start dockerd in that case.
Considering that one usually sets `skip_download: true` to speed up the
implicit "get" of the "put" operation, i think it's a meaningful
optimization.