Improve 'digest' detection
Although the `tail -1` works in current output format, it may be a
fragile detection.

We add some functionality around signing the image with the use of
`notary`, which essentially produces 2 additional lines at the end of
the output confirming the successful signature.

It is our belief, that using a simple `grep 'digest'` will be capable of
obtaining the relevant line from the output, regardless its position.

Coincidentally, the `awk` parsing remains the same.

Signed-off-by: Rafal Proszowski <paroxp@gmail.com>