Fix Check implementation to properly use HEAD requests
Fixes #311

This changes the check flow to call headDigest instead of fetchDigest directly.
If headDigest fails to get the digest it will fall back to using fetchDigest.

Tested in the following scenarios:
- Upstream to dockerhub with library and user repos
- Against private nexus registry with pull-through cache both directly and with registry_mirror
- Against the repo called out in #307 which spawned the initial change.

This fix is required for us to upgrade to 6.7.x as the nexus pull-through-cache works with HEAD requests and not GET requests for check
(problem for another day) as well as important for the Dockerhub ratelimiting on Nov 2

Signed-off-by: David Dobmeier <david.dobmeier@dcsg.com>